+++ pg_data_etl/__init__.py
@@ -1,4 +1,4 @@
 from .configuration_manager import configurations as connections
 import pg_data_etl.database.helpers as helpers
 import pg_data_etl.database.actions as actions
-from .database.db import Database
+from .database.Database import Database
